Rewrite thing flag handling
Thing flags are now stored as a set of flag keys (mnemonics) instead of
as bitfields. This allows any number of flags to be added even if they
go beyond 32 bits. Compatibility with vanilla patches is kept by using
the flag index when filtering things flags. Flag aliases are now also
supported using this system.